// api/src/@types/get_querystring.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface GetQueryStringSchemaInterface {
  offset?: number;
  limit?: number;
  attributes?: ("id" | "titre" | "image" | "description" | "createdAt" | "updatedAt")[];
  order?:
    | []
    | ["id" | "titre" | "image" | "description" | "createdAt" | "updatedAt"]
    | ["id" | "titre" | "image" | "description" | "createdAt" | "updatedAt", "ASC" | "DESC"][];
  where?: {
    id?: number;
    titre?: string;
    image?: string;
    desciprtion?: string;
    createdAt?: string;
    updatedAt?: string;
    [k: string]: unknown;
  };
}